docmaker/requirements.txt
python-docx>=1.1.0
pandas>=2.0.0
streamlit>=1.30.0
openpyxl>=3.1.0
xlrd>=2.0.1
